Almost complete login implementation.
At the current state when starting the server a user is prompted to set the
duration of the timer for the lobby (according to requirements).

After that up to five clients are logged in with their respective tech,
which every player is prompted to set upon starting the client.

The client will try to log in into the lobby, a database registers/verifies
the credentials of the player and (should) return the port on which the
match will be played.

Still to implement:

Thread that checks the connection (alive/disconnected) of a player and removes
it from the lobby

Methods for returning the "game" port and managing the second game connection.